backend/pyproject.toml
[tool.poetry]
name = "kubecel-backend"
version = "0.1.0"
description = ""
authors = ["Remco Kersten"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = {extras = ["redis"], version = "^5.4.0"}
mysql-connector-python = "^9.0.0"
fastapi = {extras = ["standard"], version = "^0.114.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
